+++ clients/fides-js/src/lib/consent-types.ts
@@ -130,19 +130,21 @@ export interface FidesInitOptions {
  * need to change.
  */
 export interface FidesGlobal extends Fides {
+  config?: FidesConfig;
   consent: NoticeConsent;
   experience?: PrivacyExperience | EmptyExperience;
-  geolocation?: UserGeolocation;
-  fides_string?: string | undefined;
-  options: FidesInitOptions;
   fides_meta: FidesJSMeta;
-  tcf_consent: TcfOtherConsent;
-  saved_consent: NoticeConsent;
-  gtm: typeof gtm;
+  fides_string?: string | undefined;
+  geolocation?: UserGeolocation;
   identity: FidesJSIdentity;
-  init: (config: FidesConfig) => Promise<void>;
   initialized: boolean;
+  options: FidesInitOptions;
+  saved_consent: NoticeConsent;
+  tcf_consent: TcfOtherConsent;
+  gtm: typeof gtm;
+  init: (config: FidesConfig) => Promise<void>;
   meta: typeof meta;
+  reinitialize: () => Promise<void>;
   shopify: typeof shopify;
   showModal: () => void;
 }